Fall back to VGA if the palette API isn't supported
I ported vesa to use the VBE service back in:

    commit 55f585a15f42ffe028ff37ea1f63543795dbf56e
    Author: Adam Jackson <ajax@redhat.com>
    Date:   Fri Sep 18 17:02:16 2009 -0400

	Use VBE palette load, not VGA banging.

I'm reasonably sure that worked on all the hardware I had handy at the
time. But it doesn't work in seabios, which means 8bpp is broken under
qemu.

We query this API early in initialization, and if it fails ->savedPal
will be NULL, so use that as the hint to fall back to VGA banging.

Signed-off-by: Adam Jackson <ajax@redhat.com>
Reviewed-by: Dave Airlie <airlied@redhat.com>